feat: add providers namespace to Client and Networks
